+++ changelog.md
@@ -6,6 +6,7 @@
 
 * Support for minecraft 1.12 (https://github.com/MysticCity/MinigamesAPI/issues/263)
 * Fixed chest handling during smart reset (https://github.com/MysticCity/MinigamesAPI/issues/230)
+* Fixed bungee support for empty servers (https://github.com/MysticCity/MinigamesAPI/issues/217)
 
 ###### MobEscape